+++ src/Commands/MenuUiHooks.php
@@ -10,11 +10,10 @@ use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\ChoiceQuestion;
 
 class MenuUiHooks extends DrushCommands
 {
-    use QuestionTrait;
-
     /** @var ModuleHandlerInterface */
     protected $moduleHandler;
     /** @var MenuParentFormSelectorInterface */
@@ -28,30 +27,6 @@ class MenuUiHooks extends DrushCommands
         $this->menuParentFormSelector = $menuParentFormSelector;
     }
 
-    /** @hook interact nodetype:create */
-    public function hookInteract(InputInterface $input, OutputInterface $output, AnnotationData $annotationData): void
-    {
-        if (!$this->isInstalled()) {
-            return;
-        }
-
-        $input->setOption(
-            'menu-available',
-            $this->input->getOption('menu-available') ?? $this->askMenus()
-        );
-
-        $menus = $this->input->getOption('menu-available');
-
-        if (empty($menus)) {
-            return;
-        }
-
-        $input->setOption(
-            'menu-default-parent',
-            $this->input->getOption('menu-default-parent') ?? $this->askDefaultParent($menus)
-        );
-    }
-
     /** @hook option nodetype:create */
     public function hookOption(Command $command, AnnotationData $annotationData): void
     {
@@ -60,7 +35,7 @@ class MenuUiHooks extends DrushCommands
         }
 
         $command->addOption(
-            'menu-available',
+            'menus-available',
             '',
             InputOption::VALUE_OPTIONAL,
             'The menus available to place links in for this content type.'
@@ -74,6 +49,27 @@ class MenuUiHooks extends DrushCommands
         );
     }
 
+    /** @hook on-event node-type-set-options */
+    public function hookSetOptions(InputInterface $input): void
+    {
+        if (!$this->isInstalled()) {
+            return;
+        }
+
+        $this->ensureOption('menus-available', [$this, 'askMenus'], true);
+        $menus = $this->input->getOption('menus-available');
+        if (is_string($menus)) {
+            $menus = explode(',', $menus);
+            $this->input->setOption('menus-available', $menus);
+        }
+
+        if ($menus === []) {
+            return;
+        }
+
+        $this->ensureOption('menu-default-parent', [$this, 'askDefaultParent'], true);
+    }
+
     /** @hook on-event nodetype-create */
     public function hookCreate(array &$values): void
     {
@@ -81,7 +77,7 @@ class MenuUiHooks extends DrushCommands
             return;
         }
 
-        $values['third_party_settings']['menu_ui']['available_menus'] = $this->input->getOption('menu-available');
+        $values['third_party_settings']['menu_ui']['available_menus'] = $this->input->getOption('menus-available');
         $values['third_party_settings']['menu_ui']['parent'] = $this->input->getOption('menu-default-parent') ?? '';
         $values['dependencies']['module'][] = 'menu_ui';
     }
@@ -89,28 +85,52 @@ class MenuUiHooks extends DrushCommands
     protected function askMenus(): array
     {
         $menus = menu_ui_get_menus();
-        $choices = ['- None -'];
+        $choices = ['_none' => '- None -'];
 
         foreach ($menus as $name => $label) {
             $label = $this->input->getOption('show-machine-names') ? $name : $label;
             $choices[$name] = $label;
         }
 
+        $question = (new ChoiceQuestion('Available menus', $choices, '_none'))
+            ->setMultiselect(true);
+
         return array_filter(
-            $this->choice('Available menus', $choices, true, 0)
+            $this->io()->askQuestion($question) ?: [],
+            function (string $value) {
+                return $value !== '_none';
+            }
         );
     }
 
-    protected function askDefaultParent(array $menus): string
+    protected function askDefaultParent(): string
     {
+        $menus = $this->input->getOption('menus-available');
         $menus = array_intersect_key(menu_ui_get_menus(), array_flip($menus));
         $options = $this->menuParentFormSelector->getParentSelectOptions('', $menus);
 
-        return $this->choice('Default parent item', $options, false, 0);
+        return $this->io()->choice('Default parent item', $options, 0);
     }
 
     protected function isInstalled(): bool
     {
         return $this->moduleHandler->moduleExists('menu_ui');
+    }
+
+    protected function ensureOption(string $name, callable $asker, bool $required): void
+    {
+        $value = $this->input->getOption($name);
+
+        if ($value === null) {
+            $value = $asker();
+        }
+
+        if ($required && $value === null) {
+            throw new \InvalidArgumentException(dt('The %optionName option is required.', [
+                '%optionName' => $name,
+            ]));
+        }
+
+        $this->input->setOption($name, $value);
     }
 }